+++ app/src/main/java/com/example/e_commerce/ui/products/models/ProductUIModel.kt
@@ -4,6 +4,6 @@ data class ProductUIModel(
     val id: String? = null,
     val name: String? = null,
     val description: String? = null,
-    val price: Float? = null,
+    val price: Double? = null,
 
 )
